Made changes to the test to make sure that the  change in logical is failed and then the tests were addressed to fix that the logic in moq tests

+++ SOLIDexperimentation/Student.cs
@@ -35,7 +35,7 @@ namespace SOLIDexperimentation
             return _studentMarksBasedOnSubjects.StudentPrecentage();
         }
 
-        public float GetTution(float percentage)
+        public float GetTution(double percentage)
         {
             return _studentFees.AllFeesIncluded(percentage);
         }
